fix issue with Subscribe returning no error
If an Istio service is behind the ingress, but is unhealthy, Subscribe
calls will not return an error. this is because the ingress actually
accepts the tcp connection, and write calls to the socket are
buffered in the kernel and no error is returned from this call either.

The only way to detect that the service is unavailable is to attempt to
read from it.

This solution will setup a wait group and the goroutine that is creating the
connection will wait on this. The reader thread will signal hthis wait group
after it has done an initial read on the socket and determined that it is OK.
The creating goroutine is then released to finish the subscribe call.

If the service is unhealthy, then the reader will close the socket and
the still release the wait group. The write to the now closed socket will fail
giving the mmd client the opportunity to properly return an error to
the caller.